requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.8.0

openai>=1.40.0

langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2

pythainlp>=5.0.0
beautifulsoup4>=4.12.0
requests>=2.32.0
pymupdf>=1.24.0
tqdm>=4.66.0

numpy>=1.26.0
pandas>=2.2.0
datasets>=2.20.0
ragas>=0.2.0
